Time: 7 ms (71.87%) | Memory: 23 MB (42.16%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.cpp
@@ -5,7 +5,7 @@ public:
         int end = *max_element(piles.begin(), piles.end());
         int res = end;
         while (start <= end) {
-            int mid = start + (end - start) / 2;
+            long long mid = start + (end - start) / 2;
             long long total = 0;
             for (size_t i = 0; i < piles.size(); ++i) {
                 total += ((long long)piles[i] + mid - 1) / mid;
